requirements.txt
# Core ML libraries
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2

# MLflow for experiment tracking
mlflow==2.5.0
boto3==1.28.25

# API and web framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Jupyter and notebooks
jupyter==1.0.0
notebook==7.0.3
ipykernel==6.25.1

# Kubeflow SDK
kfp==2.0.1
kfp-server-api==2.0.1

# Kubernetes client
kubernetes==27.2.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
click==8.1.7

# Monitoring and logging
prometheus-client==0.17.1
structlog==23.1.0

# Data processing
scipy==1.11.2
openpyxl==3.1.2

# Docker and containerization
docker==6.1.3

# Development tools
black==23.7.0
flake8==6.0.0
pre-commit==3.4.0